fix navigation
