+++ src/client/components/login/index.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
+import Navigation from '../navigation'
 
 export default class Login extends React.Component {
     constructor(props) {
@@ -86,20 +87,10 @@ export default class Login extends React.Component {
         const { auth, form, msg } = this.state
         return (
             <div>
-                <nav className="navbar has-shadow">
-                    <div className="container">
-                        <nav className="tabs is-large">
-                            <Link className="navbar-item is-tab" to="/">產品列表</Link>
-                            {!auth &&<Link className="navbar-item is-tab is-active" to="/login">會員登入</Link>}
-                            <Link className="navbar-item is-tab" to="/register">會員註冊</Link>
-                            <Link className="navbar-item is-tab" to="/contact">聯絡我們</Link>
-                            {auth && <Link className="navbar-item is-tab" to="/order">訂購清單</Link>}
-                        </nav>
-                    </div>
-                </nav>
+                <Navigation tab="login"/>
                 <div className="container" style={style.container}>
                     <div className="columns">
-                        <div className="column is-4">
+                        <div className="column is-6 is-offset-3">
                             <div className="box is-4" style={style.box}>
                                 <div className="field">
                                     <label className="label">電子郵件</label>
