+++ src/client/components/product/index.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
+import Navigation from '../navigation'
 import { connect } from 'react-redux'
 import { login_user } from '../../actions'
 
@@ -15,17 +16,7 @@ class Product extends React.Component {
         const { auth } = this.state
         return(          
             <div>
-                <nav className="navbar has-shadow">
-                    <div className="container">
-                        <nav className="tabs is-large">
-                            <Link className="navbar-item is-tab is-active" to="/">產品列表</Link>
-                            {!auth &&<Link className="navbar-item is-tab" to="/login">會員登入</Link>}
-                            <Link className="navbar-item is-tab" to="/register">會員註冊</Link>
-                            <Link className="navbar-item is-tab" to="/contact">聯絡我們</Link>
-                            {auth && <Link className="navbar-item is-tab" to="/order">訂購清單</Link>}
-                        </nav>
-                    </div>
-                </nav>
+                <Navigation tab="product"/>
                 <div className="container" style={style.container}>
                     <h1>this product</h1>
                 </div>
