+++ src/client/containers/contact/index.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 import config from '../../config'
+import Navigation from '../../components/navigation'
 
 export default class Contact extends React.Component {
     constructor(props) {
@@ -14,17 +15,7 @@ export default class Contact extends React.Component {
         const { auth } = this.state
         return( 
             <div>
-                <nav className="navbar has-shadow">
-                    <div className="container">
-                        <nav className="tabs is-large">
-                            <Link className="navbar-item is-tab" to="/">產品列表</Link>
-                            {!auth &&<Link className="navbar-item is-tab" to="/login">會員登入</Link>}
-                            <Link className="navbar-item is-tab" to="/register">會員註冊</Link>
-                            <Link className="navbar-item is-tab is-active" to="/contact">聯絡我們</Link>
-                            {auth && <Link className="navbar-item is-tab" to="/order">訂購清單</Link>}
-                        </nav>
-                    </div>
-                </nav>
+                <Navigation tab="contact"/>
                 <div className="container" style={style.container}>
                     <div>
                         <h2 className="title is-2">{contact.name}</h2>
