+++ src/client/containers/logo/index.js
@@ -2,12 +2,19 @@ import React from 'react'
 
 export default class Logo extends React.Component {
     render() {
-        return(     
-            <div className="" style={style.box}>       
-                <h1 className="title">
-                    This is LOGO!!!
-                </h1>
-            </div>
+        return(    
+            <section className="hero is-light">
+                <div className="hero-head">
+                    <div className="container">
+                        <div className="" style={style.box}>       
+                            <h1 className="title">
+                                This is LOGO!!!
+                            </h1>
+                        </div>
+                    </div>
+                </div>
+            </section> 
+            
         )
     }
 }
